# FAP/IOT & AI over Trusted Zones/implementation/ai-insight-service/docs/openapi.yaml
openapi: 3.0.3
info:
  title: FACIS AI Insight Service
  description: |
    REST API for AI-powered insight generation from energy and IoT datasets.
    This specification is the source of truth served at `/docs` and `/openapi.json`.
    For Trino authentication, the service automatically enables Keycloak OIDC
    password-flow token exchange when `AI_INSIGHT_TRINO__OIDC_TOKEN_URL` is set.
  version: 0.1.0
  contact:
    name: ATLAS IoT Lab GmbH
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/insights/anomaly-report:
    post:
      operationId: generateAnomalyReport
      tags: [insights]
      summary: Generate anomaly report
      description: |
        Query `gold.net_grid_hourly` for the requested window, compare
        consumption vs generation and cost behavior, detect robust statistical
        outliers (MAD-based robust z-score), and return structured context for LLM use.

        Runtime behavior:
        - OIDC disabled when `AI_INSIGHT_TRINO__OIDC_TOKEN_URL` is empty
        - OIDC password-flow enabled when `AI_INSIGHT_TRINO__OIDC_TOKEN_URL` is set
        - Trino TLS verification uses `AI_INSIGHT_TRINO__VERIFY`
        - Keycloak TLS verification uses `AI_INSIGHT_TRINO__OIDC_VERIFY`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NetGridOutlierRequest"
            example:
              start_ts: "2026-03-01T00:00:00Z"
              end_ts: "2026-03-08T00:00:00Z"
              timezone: "UTC"
              robust_z_threshold: 3.5
              include_data: false
      responses:
        "200":
          description: Governed AI insight generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsightResponse"
              example:
                context:
                  window:
                    start_ts: "2026-03-01T00:00:00+00:00"
                    end_ts: "2026-03-08T00:00:00+00:00"
                    timezone: "UTC"
                    rows_analyzed: 168
                  baseline_stats: []
                  outlier_events: []
                  cost_anomalies: []
                  narrative_hints:
                    - No robust outliers detected for the selected metrics and window.
                  summary:
                    total_outliers: 0
                    outliers_by_metric: {}
                    selected_metrics: []
        "400":
          description: Invalid input parameters (for example, start_ts >= end_ts)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: start_ts must be earlier than end_ts
        "403":
          description: Request is not authorized by policy controls
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: Agreement is not authorized
        "429":
          description: Agreement exceeded configured request rate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: Agreement rate limit exceeded
        "502":
          description: >
            Upstream query or processing failure (for example TLS trust issues,
            OIDC token exchange failure, invalid credentials, missing catalog, or Trino errors)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: Upstream Trino request timed out
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
  /api/v1/insights/city-status:
    post:
      operationId: generateCityStatus
      tags: [insights]
      summary: Generate city status insight
      description: |
        Query `gold.event_impact_daily` and `gold.streetlight_zone_hourly`, align by
        `zone_id` and date (`event_date = CAST(hour AS DATE)`), then generate
        event-to-infrastructure response patterns for LLM context.

        Statistical approach (hybrid):
        - Robust baseline shift with median and MAD
        - Lag-window evidence over `0-6h`, `6-24h`, and `24-48h`
        - Composite confidence score for each event/zone pattern

        Formulas:
        - `MAD = median(|x_i - median(x)|)`
        - `robust_z = (x - median) / (1.4826 * MAD)`
        - `shift_pct = ((x_event - x_baseline) / |x_baseline|) * 100`
        - `response_score = 0.45 * anomaly_strength + 0.35 * lag_strength + 0.20 * severity_or_activity`

        Example:
        - Baseline zone power median = 50W, event-window average = 125W
        - `shift_pct = ((125 - 50) / 50) * 100 = 150%`
        - Strong robust-z signal and early lag evidence lead to high confidence.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SmartCityCorrelationRequest"
            example:
              start_ts: "2026-03-01T00:00:00Z"
              end_ts: "2026-03-08T00:00:00Z"
              timezone: "UTC"
              include_data: false
      responses:
        "200":
          description: Governed AI city-status insight generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsightResponse"
              example:
                context:
                  window:
                    start_ts: "2026-03-01T00:00:00+00:00"
                    end_ts: "2026-03-08T00:00:00+00:00"
                    timezone: "UTC"
                    rows_analyzed: 192
                  event_response_patterns:
                    - event_date: "2026-03-03"
                      zone_id: "zone-7"
                      event_type: "accident"
                      event_count: 2
                      avg_severity: 2.5
                      active_count: 2
                      hours_analyzed: 24
                      lag_bucket_counts: { "0-6h": 5, "6-24h": 19, "24-48h": 0 }
                      baseline:
                        power_median: 50.0
                        dimming_median: 31.0
                      response:
                        event_avg_power_w: 125.0
                        event_avg_dimming_pct: 73.0
                        power_shift_pct: 150.0
                        dimming_shift_pct: 135.4839
                        power_shift_robust_z: 4.7212
                        dimming_shift_robust_z: 3.9981
                      response_score: 0.84
                      confidence: high
                  lag_distribution: { "0-6h": 5, "6-24h": 19, "24-48h": 0 }
                  zone_response_summary:
                    - zone_id: "zone-7"
                      patterns: 1
                      high_confidence_patterns: 1
                      avg_response_score: 0.84
                  high_confidence_links:
                    - event_date: "2026-03-03"
                      zone_id: "zone-7"
                      event_type: "accident"
                      response_score: 0.84
                  narrative_hints:
                    - Detected event-to-streetlight response patterns using hybrid baseline and lag analysis.
                  summary:
                    total_patterns: 1
                    high_confidence_links: 1
                    confidence_distribution: { high: 1, medium: 0, low: 0 }
        "400":
          description: Invalid input parameters (for example, start_ts >= end_ts)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: start_ts must be earlier than end_ts
        "403":
          description: Request is not authorized by policy controls
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: Missing required role
        "429":
          description: Agreement exceeded configured request rate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: Agreement rate limit exceeded
        "502":
          description: >
            Upstream query or processing failure (for example TLS trust issues,
            OIDC token exchange failure, invalid credentials, missing catalog, or Trino errors)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: Upstream Trino request timed out
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
  /api/v1/insights/energy-summary:
    post:
      operationId: generateEnergySummary
      tags: [insights]
      summary: Generate energy summary insight
      description: |
        Query `gold.net_grid_hourly`, `gold.weather_hourly`, `gold.energy_cost_daily`,
        and `gold.pv_self_consumption_daily` to compute trend/seasonality metrics
        and produce deterministic next-24h hourly forecast context.

        Statistical approach:
        - `MA_k(t) = (1/k) * sum(x_{t-i}, i=0..k-1)`
        - `slope = (x_t - x_{t-n}) / n`
        - `S_h = mean(x | hour=h)`
        - `x_hat(t+1) = level_recent + alpha * (S_hour - seasonality_mean)`

        Parameter behavior:
        - `forecast_alpha` controls how much hourly seasonality shifts each forecast point.
          - Lower (`0.2-0.4`): smoother and more conservative forecast.
          - Higher (`0.7-1.0`): more reactive to hourly seasonal patterns.
          - Formula impact: higher alpha increases `alpha * (S_hour - seasonality_mean)`.
        - `trend_epsilon` controls trend classification sensitivity.
          - Lower (`0.005-0.02`): easier to classify as `up`/`down`.
          - Higher (`0.03-0.08`): more values become `stable`.
          - Used against normalized slope to decide `up`, `down`, or `stable`.

        Example:
        - `level_recent = 42.0`, hourly seasonal uplift `= 3.0`, `alpha = 0.6`
        - `x_hat = 42.0 + 0.6 * 3.0 = 43.8`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnergyTrendForecastRequest"
            example:
              start_ts: "2026-03-01T00:00:00Z"
              end_ts: "2026-03-08T00:00:00Z"
              timezone: "UTC"
              forecast_alpha: 0.6
              trend_epsilon: 0.02
              daily_overview_strategy: "strict_daily"
              include_data: false
      responses:
        "200":
          description: Governed AI energy summary generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsightResponse"
              example:
                context:
                  window:
                    start_ts: "2026-03-01T00:00:00+00:00"
                    end_ts: "2026-03-08T00:00:00+00:00"
                    timezone: "UTC"
                    rows_analyzed: 168
                  trend_signals:
                    net_grid_kw:
                      direction: up
                      slope_per_step: 0.18
                  moving_averages:
                    net_grid_kw:
                      ma_6h: 41.3
                      ma_24h: 38.7
                      latest_value: 42.1
                  seasonality_patterns:
                    net_grid_kw:
                      "0": 35.2
                      "1": 34.8
                  forecast_24h:
                    - timestamp: "2026-03-08T01:00:00+00:00"
                      avg_consumption_kw: 52.1
                      net_grid_kw: 43.8
                      estimated_hourly_cost_eur: 5.2
                  daily_overview:
                    daily_cost_points: 7
                    daily_pv_points: 7
                    consumption_trend_daily: up
                    self_consumption_trend_daily: stable
                    source: daily_views
                  data_availability:
                    hourly_net_grid_weather:
                      count: 168
                      first: "2026-03-01T00:00:00+00:00"
                      last: "2026-03-07T23:00:00+00:00"
                    daily_cost:
                      count: 7
                      first: "2026-03-01"
                      last: "2026-03-07"
                    daily_pv_self_consumption:
                      count: 7
                      first: "2026-03-01"
                      last: "2026-03-07"
                  narrative_hints:
                    - Recent history coverage is sufficient for deterministic baseline forecasting.
                  summary:
                    forecast_points: 24
                    tracked_metrics:
                      - avg_consumption_kw
                      - estimated_hourly_cost_eur
                      - net_grid_kw
                    daily_cost_points: 7
                    daily_pv_points: 7
        "400":
          description: Invalid input parameters (for example, start_ts >= end_ts)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: start_ts must be earlier than end_ts
        "403":
          description: Request is not authorized by policy controls
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: Asset is not authorized
        "429":
          description: Agreement exceeded configured request rate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: Agreement rate limit exceeded
        "502":
          description: >
            Upstream query or processing failure (for example TLS trust issues,
            OIDC token exchange failure, invalid credentials, missing catalog, or Trino errors)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: Upstream Trino request timed out
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
...
  /api/ai/outputs/{output_id}:
    get:
      operationId: getAiOutput
      tags: [insights]
      summary: Get AI output by id
      description: |
        Retrieve a previously generated AI output entity by its identifier.

        This endpoint returns the persisted output record including:
        - original input context sent to inference (`input_data`)
        - model metadata (`llm_model`, `timestamp`)
        - raw model text (`output_text`)
        - structured JSON used by dashboard consumers (`structured_output`)

        Use this endpoint when dashboards or downstream services need to
        inspect/replay a specific generated insight without recomputing it.
      parameters:
        - in: path
          name: output_id
          required: true
          schema:
            type: string
          description: Unique AI output identifier.
      responses:
        "200":
          description: Stored AI output entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIOutputEntityResponse"
        "404":
          description: AI output id not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    NetGridOutlierRequest:
      type: object
      required: [start_ts, end_ts]
      properties:
        start_ts:
          type: string
          format: date-time
          description: >
            Inclusive start timestamp (ISO 8601) for the analysis window.
            Earlier start values increase historical coverage for moving averages,
            trend direction, and seasonality estimation.
          example: "2026-03-01T00:00:00Z"
        end_ts:
          type: string
          format: date-time
          description: >
            Exclusive end timestamp (ISO 8601) for the analysis window.
            Must be later than `start_ts`; larger windows generally produce more stable
            summaries and forecast context.
          example: "2026-03-08T00:00:00Z"
        timezone:
          type: string
          description: >
            Timezone label used in generated context and narrative rendering.
            Influences how temporal patterns (for example peak hours) are interpreted in output text.
          default: UTC
          example: UTC
        robust_z_threshold:
          type: number
          format: float
          description: MAD-based robust z-score sensitivity threshold (> 0)
          minimum: 0
          exclusiveMinimum: true
          default: 3.5
          example: 3.5
        include_data:
          type: boolean
          description: Include analytics context data in the response payload as `data`.
          default: false
    SmartCityCorrelationRequest:
      type: object
      required: [start_ts, end_ts]
      properties:
        start_ts:
          type: string
          format: date-time
          description: Inclusive start timestamp (ISO 8601)
          example: "2026-03-01T00:00:00Z"
        end_ts:
          type: string
          format: date-time
          description: Exclusive end timestamp (ISO 8601)
          example: "2026-03-08T00:00:00Z"
        timezone:
          type: string
          description: Timezone label used in generated context
          default: UTC
          example: UTC
        include_data:
          type: boolean
          description: Include analytics context data in the response payload as `data`.
          default: false
    EnergyTrendForecastRequest:
      type: object
      required: [start_ts, end_ts]
      properties:
        start_ts:
          type: string
          format: date-time
          description: Inclusive start timestamp (ISO 8601)
          example: "2026-03-01T00:00:00Z"
        end_ts:
          type: string
          format: date-time
          description: Exclusive end timestamp (ISO 8601)
          example: "2026-03-08T00:00:00Z"
        timezone:
          type: string
          description: Timezone label used in generated context
          default: UTC
          example: UTC
        forecast_alpha:
          type: number
          format: float
          description: >
            Forecast blending factor in `(0,1]`. Controls seasonal adjustment strength in
            `x_hat(t+1) = level_recent + alpha * (S_hour - seasonality_mean)`.
            Lower values smooth the forecast and reduce hourly oscillation impact.
            Higher values increase responsiveness to hour-of-day seasonality.
          minimum: 0
          exclusiveMinimum: true
          maximum: 1
          default: 0.6
          example: 0.6
        trend_epsilon:
          type: number
          format: float
          description: >
            Normalized slope threshold for trend classification.
            If `normalized_slope > trend_epsilon` => `up`;
            if `< -trend_epsilon` => `down`; otherwise `stable`.
            Lower values are more sensitive and produce more `up/down` classifications;
            higher values are stricter and classify more windows as `stable`.
          minimum: 0
          default: 0.02
          example: 0.02
        daily_overview_strategy:
          type: string
          description: >
            Strategy for `daily_overview` construction:
            `strict_daily` uses only daily Gold views and may return sparse/zero daily points
            when daily views are missing.
            `fallback_hourly` estimates daily-overview signals from hourly series when daily views are empty,
            improving resilience at the cost of less direct daily aggregation fidelity.
          enum: [strict_daily, fallback_hourly]
          default: strict_daily
          example: fallback_hourly
        include_data:
          type: boolean
          description: Include analytics context data in the response payload as `data`.
          default: false
...
    InsightResponse:
      type: object
      required: [insight_type, summary, key_findings, recommendations, metadata]
      properties:
        insight_type:
          type: string
          enum: [energy-summary, anomaly-report, city-status]
        summary:
          type: string
        key_findings:
          type: array
          items:
            type: string
        recommendations:
          type: array
          items:
            type: string
        metadata:
          $ref: "#/components/schemas/InsightMetadata"
        data:
          type: object
          nullable: true
          additionalProperties: true
          description: >
            Optional analytics context data. Returned only when request payload sets `include_data=true`.
    AIOutputEntityResponse:
      type: object
      required:
        [
          id,
          insight_type,
          agreement_id,
          asset_id,
          input_data,
          llm_model,
          output_text,
          structured_output,
          timestamp,
        ]
      properties:
        id:
          type: string
          format: uuid
        insight_type:
          type: string
        agreement_id:
          type: string
        asset_id:
          type: string
        input_data:
          type: object
          additionalProperties: true
        llm_model:
          type: string
        output_text:
          type: string
        structured_output:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          example: start_ts must be earlier than end_ts
    ValidationErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: array
          items:
            $ref: "#/components/schemas/ValidationErrorItem"
    ValidationErrorItem:
      type: object
      required: [loc, msg, type]
      properties:
        loc:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
        msg:
          type: string
        type:
          type: string